feat: change type to uint64

+++ test/dShareCompat.t.sol
@@ -260,7 +260,7 @@ contract dShareCompatTest is Test {
             to: to,
             paymentToken: _paymentToken,
             data: data,
-            deadline: block.timestamp + 30 days,
+            deadline: uint64(block.timestamp + 30 days),
             nonce: nonce
         });
